requirements.txt
astroid==2.2.5
autopep8==1.4.3
dj-database-url==0.5.0
Django==2.1.7
django-extensions==2.1.6
djangorestframework==3.9.2
gunicorn==19.9.0
isort==4.3.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
model-mommy==1.6.0
pep8==1.7.1
psycopg2==2.8.2
pycodestyle==2.5.0
pylint==2.3.1
pylint-django==2.0.6
pylint-plugin-utils==0.5
pytz==2018.9
six==1.12.0
typed-ast==1.3.1
whitenoise==4.1.2
wrapt==1.11.1
